% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{active_neighbours}
\alias{active_neighbours}
\title{Finds active neighbours by finding intersect between active neighbours and
possible current neighbours.}
\usage{
active_neighbours(x)
}
\arguments{
\item{x}{a RebelTrackDataFrameGid object}
}
\description{
Finds active neighbours by finding intersect between active neighbours and
possible current neighbours.
}
\keyword{internal}
