% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{add_neighbour_column}
\alias{add_neighbour_column}
\title{Adds column of nearest neigbours as list to dataframe - this differs from R best practice
but is common as an adjacency list in network packages - this is the most efficient representation
considering the structure of the adjacency.}
\usage{
add_neighbour_column(x, order = 1)
}
\arguments{
\item{x}{rebeltrack_dataframe_gid}

\item{order}{order of the spatial lag to apply}
}
\description{
Adds column of nearest neigbours as list to dataframe - this differs from R best practice
but is common as an adjacency list in network packages - this is the most efficient representation
considering the structure of the adjacency.
}
\keyword{internal}
