% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_dataframe.R
\name{as.data.frame.RebelTrackDataFrame}
\alias{as.data.frame.RebelTrackDataFrame}
\title{Coarce to data.frame}
\usage{
\method{as.data.frame}{RebelTrackDataFrame}(from)
}
\arguments{
\item{from}{A RebelTrackDataFrame object}
}
\description{
Coarce a RebelTrackDataFrame object to a data.frame.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

df <- as.data.frame(data)
}
}
