% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_dataframe.R
\name{as.data.frame.RebelTrackDataFrameGid}
\alias{as.data.frame.RebelTrackDataFrameGid}
\title{Coerce to data.frame}
\usage{
\method{as.data.frame}{RebelTrackDataFrameGid}(from)
}
\arguments{
\item{from}{A RebelTrackDataFrameGid object}
}
\description{
Coerce a RebelTrackDataFrame object to a data.frame.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe_gid(dataset,
                             period = "month",
                             balanced = TRUE)

df <- as.data.frame(data)
}
}
