% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{coords_to_gid}
\alias{coords_to_gid}
\title{Return monotonic prio grid gid code for pair of matrix coordinates}
\usage{
coords_to_gid(coords)
}
\arguments{
\item{coords}{- vector of matrix coordinates}
}
\description{
Return monotonic prio grid gid code for pair of matrix coordinates
}
\keyword{internal}
