% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{enforce_periodic_boundaries}
\alias{enforce_periodic_boundaries}
\title{Enforces periodic boundary conditions from mapping a spherical earth onto a grid.}
\usage{
enforce_periodic_boundaries(gid)
}
\arguments{
\item{gid}{- vector of matrix coordinates}
}
\description{
Enforces periodic boundary conditions from mapping a spherical earth onto a grid.
}
\keyword{internal}
