% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{full_adj_matrix}
\alias{full_adj_matrix}
\title{Produces sparse adjacency matrix for priogrid gids}
\usage{
full_adj_matrix(x)
}
\arguments{
\item{x}{A RebeltrackDataframeGid object}
}
\value{
A slam package simple_sparse_array object of adjecency between active priogrids
}
\description{
Produces adjacency matrix for gid grouped objects. produces spatial lags based
on present and active objects. The provided dataframe should have been given additional
columns with active_neighbours
}
\keyword{internal}
