% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_utils.R
\name{get_var_by_side}
\alias{get_var_by_side}
\title{Format variable name for side A or B}
\usage{
get_var_by_side(pattern, side)
}
\arguments{
\item{pattern}{Format string}

\item{side}{Side A/B or 1/2}
}
\description{
Format variable name for side A or B
}
\keyword{internal}
