% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{gid_to_coords}
\alias{gid_to_coords}
\title{return matrix coordinates for monotonic prigrid gid code}
\usage{
gid_to_coords(gid)
}
\arguments{
\item{gid}{gid number of prio grid cell}
}
\description{
return matrix coordinates for monotonic prigrid gid code
}
\keyword{internal}
