% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\docType{data}
\name{GROUP_BY_CONFLICT}
\alias{GROUP_BY_CONFLICT}
\title{GROUP_BY_CONFLICT}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
GROUP_BY_CONFLICT
}
\description{
Group by conflict
}
\keyword{datasets}
