% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{lag_neighbours}
\alias{lag_neighbours}
\title{Finds neigbours in square grid surrounding prio grid cell.
Lagged square is of side length 2*order +1
Returns list of neighbouring grid cells.}
\usage{
lag_neighbours(gid, order)
}
\arguments{
\item{gid}{- prio grid cell to lag surrounding area}

\item{order}{- controls the degree to which the spatial lag extends. corresponds to the radius of the lag}
}
\description{
Note that the grid the lag is calculated for is NOT considered as part of this lag.
I.E there are NO SELF CONNECTIONS IN THE RESULTING ADJACENCY MATRIX
}
\keyword{internal}
