% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_utils.R
\name{lead_lag}
\alias{lead_lag}
\title{Calculate lead/lag}
\usage{
lead_lag(x, n, order_by)
}
\arguments{
\item{x}{A vector of values}

\item{n}{Positive or negative integer giving the number of positions to lead or lag by}

\item{order_by}{Override the default ordering to use another vector}
}
\description{
Calculate lead/lag
}
\keyword{internal}
