% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{periodic_boundary_conditions}
\alias{periodic_boundary_conditions}
\title{Periodic boundary conditions from mapping a spherical earth onto a grid. takes
coordinates and returns a valid coordinate.}
\usage{
periodic_boundary_conditions(coords)
}
\arguments{
\item{gid}{- vector of matrix coordinates}
}
\description{
Periodic boundary conditions from mapping a spherical earth onto a grid. takes
coordinates and returns a valid coordinate.
}
\keyword{internal}
