% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_actor_name.R
\name{rebeltrack_actor_name}
\alias{rebeltrack_actor_name}
\title{Actor Name}
\usage{
rebeltrack_actor_name(x, var, long = FALSE)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrame} object}

\item{var}{Name of the variable to create}

\item{long}{Whether to get the long or short name}
}
\description{
Get the actor name for each \code{actor} in the \link{rebeltrack_dataframe}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_actor_name(data, actor_name)
}
}
