% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_adjacency_matrix.R
\name{rebeltrack_adjacency_matrix}
\alias{rebeltrack_adjacency_matrix}
\title{Adjacency Matrix}
\usage{
rebeltrack_adjacency_matrix(x, by = GROUP_BY_CONFLICT)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrame} object}

\item{by}{one of GROUP_BY_CONFLICT or GROUP_BY_GOVERNMENT}
}
\description{
Calculate the adjacency matrix fo each \code{actor} in the \link{rebeltrack_dataframe}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

adjacency_matrix <- rebeltrack_adjacency_matrix(data)
}
}
