% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_casualty_count.R
\name{rebeltrack_casualty_count}
\alias{rebeltrack_casualty_count}
\title{Calculate Casualties}
\usage{
rebeltrack_casualty_count(x, side, var, fill = 0, lag = 0, weight = NULL)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrame} object}

\item{side}{One of \code{SIDE_A}/\code{SIDE_B}, \code{"A"}/\code{"B"},
\code{"a"}/\code{"b"}, or \code{1}/\code{2}}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}
}
\description{
Calculate the number of casualties grouped by \code{actor} and \code{period}
specified when the \code{RebelTrackDataFrame} was created with
\link{rebeltrack_dataframe}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_casualty_count(data, SIDE_A, casualty_count)
}
}
