% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_clean_up.R
\name{rebeltrack_clean_up_gid}
\alias{rebeltrack_clean_up_gid}
\title{Removes temporary columns added to calculate adjacency matrix.}
\usage{
rebeltrack_clean_up_gid(x)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrameGid} object}
}
\description{
Removes temporary columns added to calculate adjacency matrix.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe_gid(dataset,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_clean_up_gid(data)
}
}
