% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_days_affected.R
\name{rebeltrack_days_affected_gid}
\alias{rebeltrack_days_affected_gid}
\title{Days Affected by Events}
\usage{
rebeltrack_days_affected_gid(x, var, fill = 0, lag = 0, weight = NULL)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrameGid} object}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply. Must be a gid weight matrix to maintain dimensionality}
}
\description{
Calculate the number of with events starting grouped by \code{priogrid gid} and
\code{period} specified when the \code{RebelTrackDataFrameGid} was reated with
\link{rebeltrack_dataframe_gid}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe_gid(dataset,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_days_affected_gid(data, days_affected)
}
}
