% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_distance_to_capital.R
\name{rebeltrack_distance_to_capital}
\alias{rebeltrack_distance_to_capital}
\title{Calculate Distance to Capital}
\usage{
rebeltrack_distance_to_capital(
  x,
  var,
  fill = NA,
  lag = 0,
  weight = NULL,
  func = mean,
  ...
)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrame} object}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}

\item{func}{Name of function to apply (default = mean)}

\item{...}{Additional arguments passed to \code{func}}
}
\description{
Calculate distance to capital of grouped by \code{actor} and \code{period}
specified when the \code{RebelTrackDataFrame} was created
with \link{rebeltrack_dataframe}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_distance_to_capital(data, capital_dist)
}
}
