% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_download.R
\name{rebeltrack_download}
\alias{rebeltrack_download}
\title{Download Datasets}
\usage{
rebeltrack_download()
}
\description{
Download all datasets to local storage
}
\examples{
library(rebeltrack)

\dontrun{
rebeltrack_download()
}
}
