% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_prio_grid.R
\name{rebeltrack_download_prio_grid}
\alias{rebeltrack_download_prio_grid}
\title{Download/compute PRIO-GRID-equivalent border and capital distance}
\usage{
rebeltrack_download_prio_grid(config)
}
\arguments{
\item{config}{The result of rebeltrack_config()}
}
\description{
Download/compute PRIO-GRID-equivalent border and capital distance
}
\keyword{internal}
