% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_event_count.R
\name{rebeltrack_event_count_gid}
\alias{rebeltrack_event_count_gid}
\title{Calculate Event Counts}
\usage{
rebeltrack_event_count_gid(x, var, fill = 0, lag = 0, weight = NULL)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrameGid} object}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}
}
\description{
Calculate event counts grouped by \code{priogrid_gid} and \code{period} specified
when the \code{RebelTrackDataFrameGid} was created with \link{rebeltrack_dataframe_gid}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe_gid(dataset,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_event_count_gid(data, event_count)
}
}
