% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_grids_affected.R
\name{rebeltrack_grids_affected}
\alias{rebeltrack_grids_affected}
\title{Grids Affected by Events}
\usage{
rebeltrack_grids_affected(x, var, fill = 0, lag = 0, weight = NULL)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrame} object}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}
}
\description{
Calculate the number of grids affected by events grouped by \code{actor} and
\code{period} specified when the \code{RebelTrackDataFrame} was created with
\link{rebeltrack_dataframe}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_grids_affected(data, grids_affected)
}
}
