% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_prio_grid.R
\name{rebeltrack_haversine_km}
\alias{rebeltrack_haversine_km}
\title{Vectorised great-circle (haversine) distance, in kilometres}
\usage{
rebeltrack_haversine_km(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1, lat1}{Coordinates of the first point (degrees)}

\item{lon2, lat2}{Coordinates of the second point (degrees)}
}
\description{
Vectorised great-circle (haversine) distance, in kilometres
}
\keyword{internal}
