% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_transnational_count.R
\name{rebeltrack_transnational_count}
\alias{rebeltrack_transnational_count}
\title{Transnational Count}
\usage{
rebeltrack_transnational_count(x, var, fill = 0, lag = 0, weight = NULL)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrame} object}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}
}
\description{
Calculate the number of countries where the \code{actor} is active in the
given \code{period}
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_transnational_count(data, transnational_ratio)
}
}
