% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{summarise_active_grids}
\alias{summarise_active_grids}
\title{Summarises the starting period and produces list of all priogrid_gid with entry
in the dataset - note that this may produce a denser matrix when balacned = TRUE.
this is handled in weighted_lag_gid}
\usage{
summarise_active_grids(x)
}
\arguments{
\item{x}{a RebelTrackDataFrameGid object}
}
\description{
Summarises the starting period and produces list of all priogrid_gid with entry
in the dataset - note that this may produce a denser matrix when balacned = TRUE.
this is handled in weighted_lag_gid
}
\keyword{internal}
