% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{total_active_grids}
\alias{total_active_grids}
\title{adds active grids for given time period as new column}
\usage{
total_active_grids(x)
}
\description{
adds active grids for given time period as new column
}
\keyword{internal}
