% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\name{UCDP_CONFLICT_INTERNAL}
\alias{UCDP_CONFLICT_INTERNAL}
\title{UCDP_CONFLICT_INTERNAL}
\usage{
UCDP_CONFLICT_INTERNAL
}
\description{
UCDP Dyadic Internal Armed Conflict (with or without intervention)
}
