% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\docType{data}
\name{UCDP_CONFLICT_INTERSTATE}
\alias{UCDP_CONFLICT_INTERSTATE}
\title{UCDP_CONFLICT_INTERSTATE}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
UCDP_CONFLICT_INTERSTATE
}
\description{
UCDP Dyadic Armed Conflict (Interstate)
}
\keyword{datasets}
