% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\docType{data}
\name{UCDP_NON_STATE}
\alias{UCDP_NON_STATE}
\title{UCDP_NON_STATE}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
UCDP_NON_STATE
}
\description{
UCDP Non-state conflict (type = 2)
}
\keyword{datasets}
