% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\docType{data}
\name{UCDP_ONE_SIDED}
\alias{UCDP_ONE_SIDED}
\title{UCDP_ONE_SIDED}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
UCDP_ONE_SIDED
}
\description{
UCDP One-sided violence (type = 3)
}
\keyword{datasets}
