% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\docType{data}
\name{UCDP_STATE_BASED}
\alias{UCDP_STATE_BASED}
\title{UCDP_STATE_BASED}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
UCDP_STATE_BASED
}
\description{
UCDP State-based conflict (type = 1)
}
\keyword{datasets}
