% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{vectorized_lag_neighbours}
\alias{vectorized_lag_neighbours}
\title{Vectorizes lag_neighbours for use in mutate functions}
\usage{
vectorized_lag_neighbours(gid, order)
}
\arguments{
\item{gid}{- prio grid cell to lag surrounding area}

\item{order}{- controls the degree to which the spatial lag extends. corresponds to the radius of the lag}
}
\description{
Vectorizes lag_neighbours for use in mutate functions
}
\keyword{internal}
