% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_utils.R
\name{weight_apply}
\alias{weight_apply}
\title{Apply weight matrix}
\usage{
weight_apply(x, w, n_actors, n_periods)
}
\arguments{
\item{x}{x A \code{RebelTrackDataFrame} object}

\item{w}{Weight matrix of NxNxT dimension}

\item{n_actors}{Number of actors in the panel}

\item{n_periods}{Number of time periods in the panel}
}
\description{
Apply weight matrix
}
\keyword{internal}
