% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_utils.R
\name{weight_apply_gid}
\alias{weight_apply_gid}
\title{Apply weight matrix for Gid Class}
\usage{
weight_apply_gid(x, w, n_actors, n_periods)
}
\arguments{
\item{x}{x A \code{RebelTrackDataFrameGid} object}

\item{w}{Weight matrix of NxNxT dimension - must be a slam simple sparse array}

\item{n_actors}{Number of actors in the panel}

\item{n_periods}{Number of time periods in the panel}
}
\description{
Apply weight matrix for Gid Class
}
\keyword{internal}
