% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_utils.R
\name{weighted_lag_gid}
\alias{weighted_lag_gid}
\title{Calculate weighted lag for Gid Dataframe}
\usage{
weighted_lag_gid(x, var, group, group_summary, fill, lag, weight)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrameGid} object}

\item{var}{Name of the variable to create}

\item{group}{Grouped dataset}

\item{group_summary}{Variable specific group summary}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}
}
\description{
Calculate weighted lag for Gid Dataframe
}
\keyword{internal}
