% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_actors_present.R
\name{rebeltrack_actors_present_gid}
\alias{rebeltrack_actors_present_gid}
\title{Actors Present per Priogrid_gid}
\usage{
rebeltrack_actors_present_gid(x, var, fill = 0, lag = 0, weight = NULL)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrameGid} object}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}
}
\description{
Calculate the number of actors active in each \code{priogrid} and
\code{period} specified when the \code{RebelTrackDataFrameGid} was created with
\link{rebeltrack_dataframe_gid}. Equivalent of rebeltrack_grids_affected
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe_gid(dataset,
                             side = SIDE_B,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_gid_actors_present(data, actors_present)
}
}
