% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_adjacency_matrix.R
\name{rebeltrack_adjacency_matrix_gid}
\alias{rebeltrack_adjacency_matrix_gid}
\title{Produces gid adjacency matrix}
\usage{
rebeltrack_adjacency_matrix_gid(x)
}
\arguments{
\item{x}{a RebelTrackDataFrameGid object}
}
\value{
A slam package simple_sparse_array object of adjecency between active priogrids
}
\description{
Produces gid adjacency matrix
}
