% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_dataframe.R
\name{rebeltrack_dataframe}
\alias{rebeltrack_dataframe}
\title{Create RebelTrackDataFrame Object}
\usage{
rebeltrack_dataframe(dataset, side, period, balanced = FALSE)
}
\arguments{
\item{dataset}{A RebelTrackDataSet object}

\item{side}{One of: {SIDE_A, SIDE_B} or {'A', 'B'} or {'a', 'b'} or {1, 2}}

\item{period}{Any date interval supported by \link{seq.Date} such as
"day(s)", "week(s)", "month(s)", "quarter(s)", or "year(s)", optionally
preceded by an integer, for example "3 months".}

\item{balanced}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, a
balanced panel is returned.}
}
\description{
Create a RebelTrackDataFrame grouped by \code{side} and \code{period}.
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_NON_STATE)

data <- rebeltrack_dataframe(dataset,
                             side = SIDE_A,
                             period = "month",
                             balanced = TRUE)
}
}
