% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_load_dataset.R
\name{rebeltrack_load_dataset}
\alias{rebeltrack_load_dataset}
\title{Load Dataset}
\usage{
rebeltrack_load_dataset(
  type = NULL,
  actor = NULL,
  conflict = NULL,
  region = NULL,
  country_gw3c = NULL,
  country_gw3n = NULL,
  country_iso3c = NULL,
  date_start = NULL,
  date_end = NULL,
  prec = 1,
  download = TRUE
)
}
\arguments{
\item{type}{Filter by type of event. One of: UCDP_STATE_BASED, UCDP_NON_STATE
or UCDP_ONE_SIDED.}

\item{actor}{Filter by numeric actor codes in either side-A or side-B}

\item{conflict}{Filter by conflict ID}

\item{region}{Filter by region. One of "Africa", "Americas", "Asia",
"Europe", "Middle East"}

\item{country_gw3c}{Filter by 3-letter Gleditsch and Ward country codes}

\item{country_gw3n}{Filter by 3-digit Gleditsch and Ward country codes}

\item{country_iso3c}{Filter by 3-letter ISO 3166-1 alpha-3 country codes}

\item{date_start}{Filter by events occcuring on or after \code{date_start}}

\item{date_end}{Filter by events occcuring on or before \code{date_end}}

\item{prec}{precision based on \code{where_prec} in \link{http://ucdp.uu.se/downloads/ged/ged181.pdf}}

\item{download}{Download datasets if necessary}
}
\description{
Create a RebelTrackDataSet object that contains all available datasets
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_NON_STATE)
}
}
