% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_gid_national_ratio.R
\name{rebeltrack_national_ratio_gid}
\alias{rebeltrack_national_ratio_gid}
\title{national Ratio}
\usage{
rebeltrack_national_ratio_gid(x, var, fill = 0, lag = 0, weight = NULL)
}
\arguments{
\item{x}{A \code{RebelTrackDataFrameGid} object}

\item{var}{Name of the variable to create}

\item{fill}{Default value when no events are observed in the specified period.}

\item{lag}{An integer giving the number of positions to lead or lag by.}

\item{weight}{A weight matrix to apply.}
}
\description{
Ratio of events in the grids's country and all other countries. Replaces
rebeltrack_transnational_ratio_gid
}
\examples{
library(rebeltrack)

\dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe_gid(dataset,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_national_ratio_gid(data, national_ratio)
}
}
