% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\name{UCDP_CONFLICT_EXTRASYSTEMIC}
\alias{UCDP_CONFLICT_EXTRASYSTEMIC}
\title{UCDP_CONFLICT_EXTRASYSTEMIC}
\usage{
UCDP_CONFLICT_EXTRASYSTEMIC
}
\description{
UCDP Dyadic Armed Conflict (Extrasystemic)
}
