% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebeltrack_const.R
\name{UCDP_CONFLICT_WITHOUT_INTERVENTION}
\alias{UCDP_CONFLICT_WITHOUT_INTERVENTION}
\title{UCDP_CONFLICT_WITHOUT_INTERVENTION}
\usage{
UCDP_CONFLICT_WITHOUT_INTERVENTION
}
\description{
UCDP Dyadic Armed Conflict (without intervention)
}
